+++ examples/example.py
@@ -1,3 +1,0 @@
-from mcpath import bedrock
-
-print(bedrock)