+++ examples/open_bedrock.py
@@ -1,0 +1,6 @@
+import mcpath
+import webbrowser
+
+url = mcpath.bedrock.get_executable()
+if url:
+    webbrowser.open(url)