+++ examples/open_education.py
@@ -1,0 +1,6 @@
+import mcpath
+import webbrowser
+
+url = mcpath.education.get_executable()
+if url:
+    webbrowser.open(url)