+++ examples/open_java.py
@@ -1,8 +1,9 @@
 import mcpath
 import os
 
-path = mcpath.java.launcher
-if os.name == "nt":  # Windows
-    os.system(f'"{path}"')  # os.startfile makes the launcher crash.
-else:  # other
-    os.system(f'open "{path}"')
+path = mcpath.java.get_launcher()
+if path:
+    if os.name == "nt":  # Windows
+        os.system(f'"{path}"')  # os.startfile makes the launcher crash.
+    else:  # other
+        os.system(f'open "{path}"')