+++ examples/open_preview.py
@@ -1,0 +1,6 @@
+import mcpath
+import webbrowser
+
+url = mcpath.preview.get_executable()
+if url:
+    webbrowser.open(url)