+++ mcpath/platforms/android/education.py
@@ -1,0 +1,12 @@
+"""
+Android Education Edition
+"""
+
+from mcpath.facades import Education
+
+
+class AndroidEducationEdition(Education): ...
+
+
+def instance():
+    return AndroidEducationEdition()