+++ mcpath/platforms/win/education.py
@@ -1,0 +1,20 @@
+"""
+Windows Education Edition
+"""
+
+from os import path
+from mcpath.facades import Education
+
+
+class WinEducationEdition(Education):
+    def _get_game_dir(self):
+        p = path.expandvars(
+            "%LOCALAPPDATA%\\Packages\\Microsoft.MinecraftEducationEdition_8wekyb3d8bbwe\\LocalState\\games\\com.mojang"
+        )
+        if path.isdir(p):
+            return p
+        return None
+
+
+def instance():
+    return WinEducationEdition()