+++ mcpath/platforms/win/preview.py
@@ -1,0 +1,20 @@
+"""
+Windows Preview Edition
+"""
+
+from os import path
+from mcpath.facades import Preview
+
+
+class WinPreviewEdition(Preview):
+    def _get_game_dir(self):
+        p = path.expandvars(
+            "%LOCALAPPDATA%\\Packages\\Microsoft.MinecraftWindowsBeta_8wekyb3d8bbwe\\LocalState\\games\\com.mojang"
+        )
+        if path.isdir(p):
+            return p
+        return None
+
+
+def instance():
+    return WinPreviewEdition()