+++ tests/test_bedrock.py
@@ -2,4 +2,13 @@ import mcpath
 
 
 def test_bedrock():
-    print(mcpath.bedrock)
+    mcpath.bedrock.get_game_dir()
+    mcpath.bedrock.get_executable()
+    mcpath.bedrock.get_worlds_dir()
+    mcpath.bedrock.get_resource_packs_dir()
+    mcpath.bedrock.get_behavior_packs_dir()
+    mcpath.bedrock.get_development_resource_packs_dir()
+    mcpath.bedrock.get_development_behavior_packs_dir()
+    mcpath.bedrock.get_development_skin_packs_dir()
+    mcpath.bedrock.get_screenshots_dir()
+    mcpath.bedrock.get_logs_dir()