+++ tests/test_education.py
@@ -2,4 +2,13 @@ import mcpath
 
 
 def test_education():
-    print(mcpath.education)
+    mcpath.education.get_game_dir()
+    mcpath.education.get_executable()
+    mcpath.education.get_worlds_dir()
+    mcpath.education.get_resource_packs_dir()
+    mcpath.education.get_behavior_packs_dir()
+    mcpath.education.get_development_resource_packs_dir()
+    mcpath.education.get_development_behavior_packs_dir()
+    mcpath.education.get_development_skin_packs_dir()
+    mcpath.education.get_screenshots_dir()
+    mcpath.education.get_logs_dir()