+++ tests/test_preview.py
@@ -2,4 +2,13 @@ import mcpath
 
 
 def test_preview():
-    print(mcpath.preview)
+    mcpath.preview.get_game_dir()
+    mcpath.preview.get_executable()
+    mcpath.preview.get_worlds_dir()
+    mcpath.preview.get_resource_packs_dir()
+    mcpath.preview.get_behavior_packs_dir()
+    mcpath.preview.get_development_resource_packs_dir()
+    mcpath.preview.get_development_behavior_packs_dir()
+    mcpath.preview.get_development_skin_packs_dir()
+    mcpath.preview.get_screenshots_dir()
+    mcpath.preview.get_logs_dir()