+++ tests/test_runtime.py
@@ -1,5 +1,0 @@
-import mcpath
-
-
-def test_preview():
-    print(mcpath.get_runtime("1.21.3"))